Fix move direction

+++ include/motor_game/Enemy.hpp
@@ -16,19 +16,19 @@ class Enemy : public Character
 
 	/// constructor with parameters
     /// param type : enemy by default
-    Enemy(const glm::vec3 &position, const unsigned int &speed, const std::string &type="Enemy"); /// constructor with parameters
+    Enemy(const glm::vec3 &position, const float &speed, const std::string &type="Enemy"); /// constructor with parameters
 
     // A VOIR SI ON GARDE : fonction appelée dans la fonction collide ???
     void killHero();
-   
+
     /// method to determine the behavior of an Enemy when the player is colliding with it
     void collide(Hero &hero);
-            
+
     /// brief method to display the value of Enemy's attributes
     void printElement() const;
 
 	/// default destructor of our Enemy
-    ~Enemy(); 
+    ~Enemy();
 
 
   protected:
